requirements.txt
langchain==0.2.14
langchain-core==0.2.33
langchain-community==0.2.12
langchain-openai==0.1.22

pydantic==2.7.4
python-dotenv==1.0.1

beautifulsoup4==4.12.3
requests==2.31.0

google-search-results==2.4.2
duckduckgo-search==6.1.0
